backend/requirements.txt
fastapi
uvicorn[standard]
aiofiles
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
celery[redis]
redis
python-socketio
aiohttp
pillow
imagehash
opencv-python-headless
numpy
torch
torchvision
open-clip-torch
faiss-cpu
python-magic
reportlab
hypothesis
pytest
pytest-asyncio
httpx
pydantic[email]
pydantic-settings
